+++ src/Facades/Concord.php
@@ -15,7 +15,11 @@ declare(strict_types=1);
 namespace Konekt\Concord\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Konekt\Concord\Hooks\HookEvent;
 
+/**
+ * @method static void hookInto(HookEvent $event, callable $callback, string|array $filter = null)
+ */
 class Concord extends Facade
 {
     /**